requirements.txt
chardet==4.0.0
matplotlib==3.4.1
pandas==1.2.0
Pillow==8.2.0
scikit_learn==0.24.0
torch==1.13.0
torchvision==0.9.1
tqdm==4.64.1
transformers==4.5.0
